pyproject.toml
[project]
name = "tele-rpi-bot"
description = "Telegram RPi Bot."
readme = "README.md"
authors = [
  { name="Nikhil Dhandre", email="nik.digitronik@live.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: Implementation :: CPython",
  "Operating System :: OS Independent",
]
dependencies = [
    "python-telegram-bot",
]
dynamic = [
    "version",
]
keywords = ["Telegram", "Raspberry Pi", "Bot"]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = ["pre-commit", "mypy"]

[project.urls]
repository = "https://github.com/digitronik/tele-rpi-bot"

[project.scripts]
tele-rpi-bot = "tele_rpi_bot:main"

[build-system]
# hatch-vcs uses setuptools_scm internally
requires = [
    "hatchling>=1.4.1",
    "hatch-vcs",
]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.targets.sdist]
include = ["tele_rpi_bot"]
[tool.hatch.build.targets.wheel]
packages = ["tele_rpi_bot"]


[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = [
    "I",    # isort
    "UP",   # pyupgrade
    "D",    # pydocstyle
]
ignore = ["D101"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
force-single-line = true
